backend/requirements.txt
fastapi==0.121.3
pydantic==2.12.4
uvicorn[standard]==0.38.0
psycopg2-binary
pytest
httpx